# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FallEvent {
  id: number;
  /** @nullable */
  location: string | null;
  /** @nullable */
  notes: string | null;
  smsStatus: string;
  smsSentCount: number;
  detectedAt: string;
  createdAt: string;
}

export interface ReportFallBody {
  location?: string;
  notes?: string;
}

export interface SmsResult {
  contactName: string;
  phone: string;
  success: boolean;
  message: string;
}

export interface ReportFallResponse {
  fall: FallEvent;
  smsResults: SmsResult[];
}

export interface FallsSummary {
  totalFalls: number;
  todayFalls: number;
  thisWeekFalls: number;
  /** @nullable */
  lastFallAt: string | null;
  totalContacts: number;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  relationship: string | null;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  phone: string;
  relationship?: string;
}