# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all fall events
 */
export const ListFallsResponseItem = zod.object({
  id: zod.number(),
  location: zod.string().nullable(),
  notes: zod.string().nullable(),
  smsStatus: zod.string(),
  smsSentCount: zod.number(),
  detectedAt: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const ListFallsResponse = zod.array(ListFallsResponseItem);

/**
 * Creates a fall event and sends SMS alerts to all emergency contacts
 * @summary Report a fall detected event
 */
export const ReportFallBody = zod.object({
  location: zod.string().optional(),
  notes: zod.string().optional(),
});

/**
 * @summary Get summary statistics for fall events
 */
export const GetFallsSummaryResponse = zod.object({
  totalFalls: zod.number(),
  todayFalls: zod.number(),
  thisWeekFalls: zod.number(),
  lastFallAt: zod.string().nullable(),
  totalContacts: zod.number(),
});

/**
 * @summary Delete a fall event
 */
export const DeleteFallParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all emergency contacts
 */
export const ListContactsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  relationship: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});
export const ListContactsResponse = zod.array(ListContactsResponseItem);

/**
 * @summary Add an emergency contact
 */
export const CreateContactBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  relationship: zod.string().optional(),
});

/**
 * @summary Remove an emergency contact
 */
export const DeleteContactParams = zod.object({
  id: zod.coerce.number(),
});