# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactBody {
  name: string;
  phone: string;
  relationship?: string;
}