# lib/api-zod/src/generated/types/fallEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface FallEvent {
  id: number;
  /** @nullable */
  location: string | null;
  /** @nullable */
  notes: string | null;
  smsStatus: string;
  smsSentCount: number;
  detectedAt: Date;
  createdAt: Date;
}