# lib/api-zod/src/generated/types/fallsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */

export interface FallsSummary {
  totalFalls: number;
  todayFalls: number;
  thisWeekFalls: number;
  /** @nullable */
  lastFallAt: string | null;
  totalContacts: number;
}