# lib/api-zod/src/generated/types/reportFallResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fall Detection Alert System API
 * OpenAPI spec version: 0.1.0
 */
import type { FallEvent } from "./fallEvent";
import type { SmsResult } from "./smsResult";

export interface ReportFallResponse {
  fall: FallEvent;
  smsResults: SmsResult[];
}